requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
reportlab==4.2.2
matplotlib==3.9.2
pillow==10.4.0